utilities/requirements.txt
boto3==1.17.88
django-extensions==3.1.3
django-filter==2.4.0
django-import-export==2.5.0
django-jet==1.0.4
django-localflavor==3.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-simple-history==3.0.0
django-storages==1.11.1
djangorestframework==3.12.4
phonenumbers==8.12.24
requests==2.25.1
boto3==1.17.88
celery==5.1.2
DateTimeRange==1.2.0
django-extensions==3.1.3
Django==3.2.4
django-3-jet==1.0.8
django-allauth==0.44.0
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-import-export==2.5.0
django-localflavor==3.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-simple-history==3.0.0
django-storages==1.11.1
djangorestframework==3.12.4
drf-nested-routers==0.93.3
drf-yasg==1.20.0
google-api-python-client==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
indian-cities==1.0.0
inflect==5.3.0
phonenumbers==8.12.24
Pillow==8.2.0
psycopg2==2.8.6
pyotp==2.6.0
redis==3.5.3
requests==2.25.1
django-taggit==1.4.0
django-multiselectfield==0.1.12
django-nested-admin~=3.3.3
